Setting up proper test envs
Alright, so I'm currently working on finishing up the ui to db codebase for the assignment. But the issue is as follows: 1) we're not using proper test envs, and 2) the database fixture is creating the initial user to create the emailAlreadyExists error, but it's not returning this, or it's not generating in the view. Let's investigate.